webrtc/requirements.txt
fastapi
uvicorn[standard]
pyserial
pygame
